Fix a lot of stuff
