+++ app/blog/page.tsx
@@ -1,20 +1,15 @@
-import { use } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
-import Layout from 'components/Layout';
-import { getPosts, BlogPost } from 'utils/posts';
+import { getPosts } from 'utils/posts';
 import { marked } from 'marked';
 import { formatDate } from 'utils/date';
 import Simpsons from 'public/images/blog/simpsons-any-key.jpg';
 
-async function getProps(): Promise<BlogPost[]> {
-  return (await getPosts('trim')).sort((a, b) => b.date.localeCompare(a.date));
-}
-
-export default function Blog() {
-  const posts = use(getProps());
+export default async function Blog() {
+  const allPosts = await getPosts('trim');
+  const posts = allPosts.sort((a, b) => b.date.localeCompare(a.date));
   return (
-    <Layout title="Blog">
+    <>
       <h1>Blog</h1>
       <Image src={Simpsons} placeholder="blur" width="720" height="420" alt="Simpsons Any Key" />
       {posts.map(({ slug, title, date, content }) => (
@@ -40,6 +35,6 @@ export default function Blog() {
           <hr />
         </article>
       ))}
-    </Layout>
+    </>
   );
 }
