+++ app/contact/page.tsx
@@ -1,10 +1,9 @@
 import Image from 'next/image';
-import Layout from 'components/Layout';
 import Kirk from 'public/images/blog/star-trek-kirk.jpg';
 
 export default function Contact() {
   return (
-    <Layout title="Contact">
+    <>
       <h1>First Contact 🖖</h1>
       <Image src={Kirk} placeholder="blur" width="720" height="360" alt="Star Trek Kirk" />
       <p>Hey thanks for reaching out! You can contact me in the following ways:</p>
@@ -36,6 +35,6 @@ export default function Contact() {
         You can also send an email to me <code>@styfle.dev</code> if you&apos;re into that kind of
         stuff.
       </p>
-    </Layout>
+    </>
   );
 }
