+++ package.json
@@ -19,7 +19,7 @@
     "gray-matter": "4.0.3",
     "highlight.js": "11.6.0",
     "marked": "4.1.1",
-    "next": "13.0.1",
+    "next": "13.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
@@ -30,7 +30,7 @@
     "@types/node": "16",
     "@types/react": "18.0.21",
     "eslint": "8.26.0",
-    "eslint-config-next": "13.0.1",
+    "eslint-config-next": "13.0.2",
     "prettier": "2.7.1",
     "typescript": "4.8.4"
   }
